wip: refactor data publisher transmittal sheet components
