+++ hmda-data-publisher/src/main/scala/hmda/publisher/query/component/PublisherComponent2020.scala
@@ -166,34 +166,6 @@ trait PublisherComponent2020 extends PGTableNameLoader {
     TableQuery(tag => new QATransmittalSheetTable(tag, tableName))
   }
 
-  class TSRepository2020Base[TsTable <: AbstractTransmittalSheetTable[TransmittalSheetEntity]](
-    val config: DatabaseConfig[JdbcProfile],
-    val table: TableQuery[TsTable]
-  ) extends TsRepository[TsTable] {
-
-    override def getId(row: TsTable): config.profile.api.Rep[Id] =
-      row.lei
-
-    def createSchema() = db.run(table.schema.create)
-    def dropSchema()   = db.run(table.schema.drop)
-
-    def insert(ts: TransmittalSheetEntity): Future[Int] =
-      db.run(table += ts)
-
-    def findByLei(lei: String): Future[Seq[TransmittalSheetEntity]] =
-      db.run(table.filter(_.lei === lei).result)
-
-    def deleteByLei(lei: String): Future[Int] =
-      db.run(table.filter(_.lei === lei).delete)
-
-    def count(): Future[Int] =
-      db.run(table.size.result)
-
-    def getAllSheets(bankIgnoreList: Array[String]): Future[Seq[TransmittalSheetEntity]] =
-      db.run(table.filterNot(_.lei.toUpperCase inSet bankIgnoreList).result)
-
-  }
-
   abstract class LarTableBase[T](tag: Tag, tableName: String) extends Table[T](tag, tableName) {
 
     def id                         = column[Int]("id")
@@ -501,7 +473,7 @@ trait PublisherComponent2020 extends PGTableNameLoader {
   }
 
   def createTransmittalSheetRepository2020(config: DatabaseConfig[JdbcProfile], p: Year2020Period) =
-    new TSRepository2020Base(config, transmittalSheetTableQuery2020(p))
+    new TsRepository[TransmittalSheetTable](config, transmittalSheetTableQuery2020(p))
 
   def createQaTransmittalSheetRepository2020(config: DatabaseConfig[JdbcProfile], p: Year2020Period)(implicit ec: ExecutionContext) =
     new QARepository.Default[TransmittalSheetEntity, QATransmittalSheetTable](config, qaTransmittalSheetTableQuery2020(p))(ec)
